Move utils/random => random
